+++ apps/@examples/modal-custom/src/hooks/use-identify-user.ts
@@ -1,30 +1,27 @@
 import { useState, useEffect } from 'react';
 
 import {
-  IdentifyApi,
-  Configuration,
-  IdentifyUserRequestBody,
+  DoptApi,
+  DoptApiClient as UsersApiClient,
 } from '@dopt/users-javascript-client';
 
 import { getUserId } from '@/utils/user';
 
-type Identifier = IdentifyUserRequestBody['identifier'];
-type Properties = IdentifyUserRequestBody['properties'];
+type Identifier = DoptApi.IdentifyUserRequestBody['identifier'];
+type Properties = DoptApi.IdentifyUserRequestBody['properties'];
 
 export function useIdentifyUser(properties: Properties) {
-  const [usersClient] = useState<IdentifyApi>(
-    new IdentifyApi(
-      new Configuration({
-        apiKey: import.meta.env.VITE_DOPT_USERS_API_KEY,
-      })
-    )
+  const [usersClient] = useState<UsersApiClient>(
+    new UsersApiClient({
+      apiKey: import.meta.env.VITE_DOPT_USERS_API_KEY,
+    })
   );
 
   const [userId, setUserId] = useState<string>();
 
   useEffect(() => {
     async function identifyUser(uid: Identifier) {
-      await usersClient.identifyUser({
+      await usersClient.users.identifyUser({
         identifier: uid,
         properties,
       });