+++ apps/@examples/tour/src/components/skeleton/index.tsx
@@ -1,19 +1,9 @@
-import { PropsWithChildren } from 'react';
-import LoadingSkeleton from 'react-loading-skeleton';
-import { skeletonClass } from './index.css';
+import { type ForwardedRef, forwardRef } from 'react';
+import './styles.css';
 
-type Props = PropsWithChildren;
-
-export function Skeleton({ children }: Props) {
-  return (
-    <div className={skeletonClass}>
-      <LoadingSkeleton
-        baseColor="#F1F3F5"
-        height={200}
-        width={200}
-        enableAnimation={false}
-      />
-      {children}
-    </div>
-  );
+function Skeleton(_props: unknown, ref?: ForwardedRef<HTMLDivElement>) {
+  return <div className="skeleton" ref={ref}></div>;
 }
+
+const Component = forwardRef(Skeleton);
+export { Component as Skeleton };