+++ apps/@examples/tour/src/pages/_app.tsx
@@ -1,11 +1,12 @@
 import { DoptProvider } from '@dopt/react';
 import { useIdentifyUser } from '@/hooks';
 
-import 'react-loading-skeleton/dist/skeleton.css';
+import { Home } from '@/pages';
 
-import { Example } from '@/pages/';
+import './_app.css';
 
 export function App() {
+  // Identify a dummy user to Dopt when the app loads
   const userId = useIdentifyUser({
     company: 'Dopt',
     role: 'admin',
@@ -16,9 +17,9 @@ export function App() {
     <DoptProvider
       apiKey={import.meta.env.VITE_DOPT_BLOCKS_API_KEY}
       userId={userId}
-      flowVersions={{ tour: 3 }}
+      flowVersions={{ 'tour-component': 1 }}
     >
-      <Example />
+      <Home />
     </DoptProvider>
   );
 }