+++ apps/@examples/tour/vite.config.ts
@@ -1,6 +1,5 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
-import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
 
 export default defineConfig({
   resolve: {
@@ -9,5 +8,5 @@ export default defineConfig({
     },
   },
   base: process.env.NODE_ENV === 'production' ? `/tour/` : '/',
-  plugins: [react(), vanillaExtractPlugin()],
+  plugins: [react()],
 });