+++ apps/@examples/checklist/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @examples/checklist
 
+## 0.1.23
+
+### Patch Changes
+
+- Updated dependencies [6eef0f305]
+  - @dopt/users-javascript-browser-client@1.0.1
+
+## 0.1.22
+
+### Patch Changes
+
+- fc6cf1251: update to use @dopt/react-checklist
+- Updated dependencies [9d0a4c592]
+  - @dopt/users-javascript-browser-client@1.0.0
+
 ## 0.1.21
 
 ### Patch Changes