+++ apps/@examples/tour/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @examples/tour
 
+## 0.1.0
+
+### Minor Changes
+
+- d3cf8e941: use @dopt/react-tour
+
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [6eef0f305]
+  - @dopt/users-javascript-browser-client@1.0.1
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [9d0a4c592]
+  - @dopt/users-javascript-browser-client@1.0.0
+
 ## 0.0.26
 
 ### Patch Changes